cap_extint: avoid signed shift overflow in cap_copy_int()
cap_copy_int() reconstructs each 32-bit capability word by shifting the
imported bytes into place. The bytes are __u8 and integer-promote to
signed int before the shift, so for the high byte "byte << 24" is
undefined behaviour whenever bit 7 is set: 0x80 << 24 exceeds INT_MAX.

This is reached through the public cap_copy_int()/cap_copy_int_check()
import path for the portable external representation, and the trigger is
an ordinary capability: CAP_SETFCAP (bit 31) sets the top byte of word 0
to 0x80. Importing legitimate capability data therefore aborts on
UBSan- or -ftrapv-hardened builds, though a normal two's-complement
build happens to compute the intended value.

Widen each byte to __u32 before shifting so the shift is performed in
an unsigned type. The round-trip is unchanged functionally (CAP_SETFCAP
is preserved); only the undefined behaviour is removed.

Found by fuzzing the import paths with clang ASan+UBSan; verified that a
CAP_SETFCAP round-trip is UBSan-clean after the change.

+++ libcap/cap_extint.c
@@ -167,14 +167,22 @@ cap_t cap_copy_int(const void *cap_ext)
 	for (blk=0; blk<(CAP_SET_SIZE/sizeof(__u32)); ++blk) {
 	    __u32 val = 0;
 
+	    /*
+	     * Each byte is a __u8 which the usual arithmetic conversions
+	     * promote to (signed) int before the shift. Shifting a value
+	     * with bit 7 set left by 24 then overflows int (e.g. a set
+	     * containing CAP_SETFCAP makes the high byte 0x80), which is
+	     * undefined behavior. Widen to __u32 so the shifts are
+	     * performed in unsigned arithmetic and are always well defined.
+	     */
 	    if (bno != blen)
 		val  = export->bytes[bno++][set];
 	    if (bno != blen)
-		val |= export->bytes[bno++][set] << 8;
+		val |= (__u32) export->bytes[bno++][set] << 8;
 	    if (bno != blen)
-		val |= export->bytes[bno++][set] << 16;
+		val |= (__u32) export->bytes[bno++][set] << 16;
 	    if (bno != blen)
-		val |= export->bytes[bno++][set] << 24;
+		val |= (__u32) export->bytes[bno++][set] << 24;
 
 	    cap_d->u[blk].flat[set] = val;
 	}
